Rows in the tables can now be clicked, also minor changes
All of the tables have been updated so that they have a OnClick event allowing the information to be edited without the edit button needed, on hover colour change shows current selection.

Elements on the tables have been centered to look better, spacing in places that make sense and capitilisation on words that needed it.